requirements.txt
pillow
fastai 
streamlit-back-camera-input